added sites endpoint to api

+++ packages/frontend/lib/_electron/authManager.js
@@ -73,7 +73,7 @@ class AuthManager {
    * @return {Object} valid session or null
    */
   async getSession() {
-    if (nebulaEnv.isDevelopment()) {
+    if (nebulaEnv.isDevelopment() && !process.env.NEBULA_ENABLE_AUTH) {
       return {
         accessToken: 'DEVACCESS',
         refreshToken: 'DEVREFRESH',
@@ -118,7 +118,7 @@ class AuthManager {
    * @return {Boolean} valid user
    */
   async checkSession() {
-    if (nebulaEnv.isDevelopment()) {
+    if (nebulaEnv.isDevelopment() && !process.env.NEBULA_ENABLE_AUTH) {
       return true;
     }
 
@@ -155,7 +155,7 @@ class AuthManager {
   }
 
   async createActiveSession() {
-    if (nebulaEnv.isDevelopment()) {
+    if (nebulaEnv.isDevelopment() && !process.env.NEBULA_ENABLE_AUTH) {
       return true;
     }
 
@@ -181,7 +181,7 @@ class AuthManager {
   }
 
   async removeActiveSession() {
-    if (nebulaEnv.isDevelopment()) {
+    if (nebulaEnv.isDevelopment() && !process.env.NEBULA_ENABLE_AUTH) {
       return true;
     }
 
@@ -211,7 +211,7 @@ class AuthManager {
    * @return {Boolean} valid attempt to clear
    */
   async clearSession() {
-    if (nebulaEnv.isDevelopment()) {
+    if (nebulaEnv.isDevelopment() && !process.env.NEBULA_ENABLE_AUTH) {
       this._store.delete('session');
       return true;
     }
@@ -242,7 +242,7 @@ class AuthManager {
    * @return {Object} valid session or errors
    */
   async createSession(key) {
-    if (nebulaEnv.isDevelopment()) {
+    if (nebulaEnv.isDevelopment() && !process.env.NEBULA_ENABLE_AUTH) {
       return {
         accessToken: 'DEVACCESS',
         refreshToken: 'DEVREFRESH',
